+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+## [0.4.0](https://github.com/chrizzFTD/naming/releases/tag/Release-0.4.0) - 2019-10-26
+### Changed
+- Renamed `drops` to `drop`
+- Renamed `compounds` to `join`
+- Added `join_sep` class property
+
 ## [0.3.0](https://github.com/chrizzFTD/naming/releases/tag/Release-0.3.0) - 2019-05-06
 ### Changed
 - `BaseName` renamed to just `Name` on `naming` module.